% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{crime_scene_report}
\alias{crime_scene_report}
\title{Crime Scene Data for dplyr Murder Mystery}
\format{
A data frame with 1228 observations of 4 variables
\describe{
  \item{date}{Date of the report}
  \item{type}{Type of crime in the report}
  \item{description}{Witness description of the crime}
  \item{city}{City that crime occured in}
 }
}
\source{
\url{https://github.com/NUKnightLab/sql-mysteries}

\url{https://github.com/veltman/clmystery}
}
\usage{
crime_scene_report
}
\description{
There has been a murder in dplyr City! All you know is that it took place on January 15, 2018 in dplyr City. All
the clues that you will need to solve the murder are in the data in this package. Use your dplyr skillz to solve
the murder mystery, have fun and no spoilers!
}
\details{
The dplyr murder mystery is a murder mystery not entirely unlike the SQL murder mystery listed in the source. The
primary difference is that we added R based easter eggs. Have fun looking for them!
}
\keyword{datasets}
