% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivers_license.R
\docType{data}
\name{drivers_license}
\alias{drivers_license}
\title{Drivers License Data for dplyr Murder Mystery}
\format{
A data frame with 10007 observations of 9 variables
\describe{
  \item{license_id}{Drivers License id; not the same as person_id in other data}
  \item{age}{Age in years}
  \item{height}{height in inches}
  \item{eye_color}{Eye color}
  \item{hair_color}{Hair color}
  \item{gender}{male or female}
  \item{plate_number}{License plate number (also has letters)}
  \item{car_make}{Make of car}
  \item{car_model}{Model of car}
 }
}
\source{
\url{https://github.com/NUKnightLab/sql-mysteries}

\url{https://github.com/veltman/clmystery}
}
\usage{
drivers_license
}
\description{
There has been a murder in dplyr City! All you know is that it took place on January 15, 2018 in dplyr City. All
the clues that you will need to solve the murder are in the data in this package. Use your dplyr skillz to solve
the murder mystery, have fun and no spoilers!
}
\details{
The dplyr murder mystery is a murder mystery not entirely unlike the SQL murder mystery listed in the source. The
primary difference is that we added R based easter eggs. Have fun looking for them!
}
\keyword{datasets}
