% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook_event.R
\docType{data}
\name{facebook_event_checkin}
\alias{facebook_event_checkin}
\title{Facebook Event Checkin Data for dplyr Murder Mystery}
\format{
A data frame with 10007 observations of 9 variables
\describe{
  \item{person_id}{Person id. Consistent across data frames}
  \item{event_id}{Id number of the event}
  \item{event_name}{Event name}
  \item{date}{Date of the event}
 }
}
\source{
\url{https://github.com/NUKnightLab/sql-mysteries}

\url{https://github.com/veltman/clmystery}
}
\usage{
facebook_event_checkin
}
\description{
There has been a murder in dplyr City! All you know is that it took place on January 15, 2018 in dplyr City. All
the clues that you will need to solve the murder are in the data in this package. Use your dplyr skillz to solve
the murder mystery, have fun and no spoilers!
}
\details{
The dplyr murder mystery is a murder mystery not entirely unlike the SQL murder mystery listed in the source. The
primary difference is that we added R based easter eggs. Have fun looking for them!
}
\keyword{datasets}
