% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit_check.R
\docType{data}
\name{get_fit_now_checkin}
\alias{get_fit_now_checkin}
\title{Gym Checkin Data for dplyr Murder Mystery}
\format{
A data frame with 2703 observations of 4 variables
\describe{
  \item{membership_id}{Member ID. Unique to Get Fit Gym}
  \item{check_in_date}{Check in Date yyyymmdd}
  \item{check_in_time}{Check in time. Format unclear.}
  \item{check_out_time}{Check out time. Format unclear.}
 }
}
\source{
\url{https://github.com/NUKnightLab/sql-mysteries}

\url{https://github.com/veltman/clmystery}
}
\usage{
get_fit_now_checkin
}
\description{
There has been a murder in dplyr City! All you know is that it took place on January 15, 2018 in dplyr City. All
the clues that you will need to solve the murder are in the data in this package. Use your dplyr skillz to solve
the murder mystery, have fun and no spoilers!
}
\details{
The dplyr murder mystery is a murder mystery not entirely unlike the SQL murder mystery listed in the source. The
primary difference is that we added R based easter eggs. Have fun looking for them!
}
\keyword{datasets}
