% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit_member.R
\docType{data}
\name{get_fit_now_member}
\alias{get_fit_now_member}
\title{Gym Membership Data for dplyr Murder Mystery}
\format{
A data frame with 185 observations of 5 variables
\describe{
  \item{id}{Gum membership ID. Consistent with `get_fit_now_checkin`}
  \item{person_id}{Person ID. Consistent across all data frames.}
  \item{name}{name of gym member}
  \item{membership_start_date}{Date of the Start date of membership yyyymmdd}
  \item{membership_status}{Status of membership. One of gold, regular or silver.}
 }
}
\source{
\url{https://github.com/NUKnightLab/sql-mysteries}

\url{https://github.com/veltman/clmystery}
}
\usage{
get_fit_now_member
}
\description{
There has been a murder in dplyr City! All you know is that it took place on January 15, 2018 in dplyr City. All
the clues that you will need to solve the murder are in the data in this package. Use your dplyr skillz to solve
the murder mystery, have fun and no spoilers!
}
\details{
The dplyr murder mystery is a murder mystery not entirely unlike the SQL murder mystery listed in the source. The
primary difference is that we added R based easter eggs. Have fun looking for them!
}
\keyword{datasets}
