% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interview.R
\docType{data}
\name{interview}
\alias{interview}
\title{Interview Data for dplyr Murder Mystery}
\format{
A data frame with 4991 observations of 2 variables
\describe{
  \item{person_id}{Person id. Consistent across data frames}
  \item{transcript}{Transcript of interview with person}
 }
}
\source{
\url{https://github.com/NUKnightLab/sql-mysteries}

\url{https://github.com/veltman/clmystery}
}
\usage{
interview
}
\description{
There has been a murder in dplyr City! All you know is that it took place on January 15, 2018 in dplyr City. All
the clues that you will need to solve the murder are in the data in this package. Use your dplyr skillz to solve
the murder mystery, have fun and no spoilers!
}
\details{
The dplyr murder mystery is a murder mystery not entirely unlike the SQL murder mystery listed in the source. The
primary difference is that we added R based easter eggs. Have fun looking for them!
}
\keyword{datasets}
